requirements.txt
beautifulsoup4==4.13.3
black==25.1.0
crispy-bootstrap4==2024.10
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
